tools: Refactor rofi-tool a bit, add a new option
Separate rendering part of the rofi-tool into separate header,
add an option `build` for showing implemented modules/rofibots.
