+++ tools/rofiTool/build.hpp
@@ -1,0 +1,3 @@
+#include <dimcli/cli.h>
+
+int build( Dim::Cli & cli );
