+++ tools/rofiTool/main.cpp
@@ -2,11 +2,13 @@
 #include <iostream>
 #include "preview.hpp"
 #include "check.hpp"
+#include "build.hpp"
 
 int main( int argc, char * argv[] ) {
     Dim::Cli cli;
     cli.command( "preview" ).action( preview );
     cli.command( "check"   ).action( check   );
+    cli.command( "build"   ).action( build   );
     try {
         return cli.exec( std::cerr, argc, argv );
     }
