+++ tools/rofiTool/preview.hpp
@@ -1,4 +1,3 @@
 #include <dimcli/cli.h>
 
 int preview( Dim::Cli & cli );
-
