+++ tools/rofiTool/rendering.hpp
@@ -1,0 +1,3 @@
+#include <rofibot.hpp>
+
+void renderConfiguration( rofi::Rofibot configuration, const std::string& configName = "unknown" );
